requirements.txt
livekit-agents[silero,turn-detector]~=1.3
livekit-plugins-noise-cancellation~=0.2
python-dotenv
onnxruntime<1.20; sys_platform == 'darwin' and platform_machine == 'x86_64'
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-cartesia
aiohttp
livekit-agents[mcp]
streamlit
livekit